% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manymodels-function.R
\name{manymodels}
\alias{manymodels}
\title{A many model function}
\usage{
manymodels(fitlist, methodnamevector)
}
\arguments{
\item{fitlist}{A list of model fit objects}

\item{methodnamevector}{A vector of corresponding model method names}
}
\value{
plot of coefficients
}
\description{
This function plots the coefficients from multiple models
}
\examples{
manymodels()
}

