% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manymodels-function.R
\name{manymodels}
\alias{manymodels}
\title{A many model function}
\usage{
manymodels(fitlist, methodnamevector)
}
\arguments{
\item{fitlist}{A list of model fit objects}

\item{methodnamevector}{A vector of corresponding model method names}
}
\value{
plot of coefficients
}
\description{
This function plots the coefficients from multiple models
}
\examples{
x1 <- rnorm(100)
x2 <- rnorm(100)
y1 <- 2 + 4*x1 + 1*x2 + rnorm(100)
y2 <- 3.5*x1 + 0.5*x2 + rnorm(100)
fit1 <- lm(y1 ~ x1 + x2)
fit2 <- lm(y2 ~ x1 + x2)
fits <- list(fit1, fit1)
methods <- c("fit1", "fit2")
}

